fix(plugin-swc): output.charset not work
