+++ packages/shared/src/minimize.ts
@@ -1,5 +1,10 @@
-import { TerserPluginOptions, NormalizedConfig } from './types';
+import { isObject } from './utils';
 import { mergeChainedOptions } from './mergeChainedOptions';
+import type {
+  NormalizedConfig,
+  TerserPluginOptions,
+  RspackBuiltinsConfig,
+} from './types';
 
 function applyRemoveConsole(
   options: TerserPluginOptions,
@@ -31,7 +36,7 @@ function applyRemoveConsole(
   return options;
 }
 
-export async function getJSMinifyOptions(config: NormalizedConfig) {
+export async function getTerserMinifyOptions(config: NormalizedConfig) {
   const DEFAULT_OPTIONS: TerserPluginOptions = {
     terserOptions: {
       mangle: {
@@ -69,3 +74,44 @@ export async function getJSMinifyOptions(config: NormalizedConfig) {
 
   return finalOptions;
 }
+
+export const getSwcMinimizerOptions = (config: NormalizedConfig) => {
+  const options: RspackBuiltinsConfig['minifyOptions'] = {};
+
+  const { removeConsole } = config.performance;
+
+  if (removeConsole === true) {
+    options.compress = {
+      ...(isObject(options.compress) ? options.compress : {}),
+      drop_console: true,
+    };
+  } else if (Array.isArray(removeConsole)) {
+    const pureFuncs = removeConsole.map((method) => `console.${method}`);
+    options.compress = {
+      ...(isObject(options.compress) ? options.compress : {}),
+      pure_funcs: pureFuncs,
+    };
+  }
+
+  options.format ||= {};
+
+  switch (config.output.legalComments) {
+    case 'inline':
+      options.format.comments = 'some';
+      options.extractComments = false;
+      break;
+    case 'linked':
+      options.extractComments = true;
+      break;
+    case 'none':
+      options.format.comments = false;
+      options.extractComments = false;
+      break;
+    default:
+      break;
+  }
+
+  options.format.asciiOnly = config.output.charset === 'ascii';
+
+  return options;
+};
